% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincom.R
\name{lincom}
\alias{lincom}
\title{Create estimates, CIs and P-values for a linear combination of two coefficients}
\usage{
lincom(model, lc, robust = FALSE, expo = TRUE, digits = 3,
  type = "HC1", level = 0.95, df = Inf)
}
\arguments{
\item{model}{The fit of a generalized linear model (lm or glm)}

\item{lc}{Character vector for the linear combination(s) set equal to null value(s) (usually zero) to test.  Written in terms of variable names.}

\item{robust}{Logical expression indicating whether robust standard errors are to be used}

\item{expo}{Logical expression indicating whether exponentiated linear combination of coefficients is requiested}

\item{digits}{Number of digits to round to in output}

\item{type}{Character string giving type of robust standard error from sandwich package}

\item{level}{Numeric confidence coefficient for CIs (as proportion)}

\item{df}{Degrees of freedom for reference t distribution (Inf means Normal distribution)}
}
\value{
A table of linear combination values or exponentiated linear combination values with CI(s)s (robust or not) and P-value(s).
}
\description{
Create estimates, CIs and P-values for a linear combination of two coefficients
}
\examples{
x1 <- rbinom(100, 2, .5)
x2 <- rbinom(100, 2, .5)
expit <- function(x) exp(x)/(1 + exp(x))
y <- rbinom(100, 1, p = expit(x1 + x2 + .1))
model <- glm(y~x1*x2, family = "binomial")
variable.names(model)
lincom(model, lc = "x1 + x1:x2 == 0")
}
