% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{robust}
\alias{robust}
\title{Computes a coefficient summary table for glms with robust standard errors and robust CIs}
\usage{
robust(model, type = c("HC1"), df = Inf, expo = TRUE, level = 0.95,
  digits = 3)
}
\arguments{
\item{model}{The fit of a generalized linear model (lm or glm)}

\item{type}{Character string giving type of robust standard error from sandwich package}

\item{df}{Degrees of freedom for reference t distribution (Inf means Normal distribution)}

\item{expo}{Logical expression indicating whether exponentiated linear combination of coefficients is requiested}

\item{level}{Numeric confidence coefficient for CIs (as proportion)}

\item{digits}{Number of digits to round to in output}
}
\value{
A table with coefficient estimates (exponentiated or not), CIs and P-values based on robust standard errors
}
\description{
Computes a coefficient summary table for glms with robust standard errors and robust CIs
}
\examples{
x <- rbinom(100, 2, .5)
expit <- function(x) exp(x)/(1 + exp(x))
y <- rbinom(100, 1, p = expit(x + .1))
model <- glm(y~x, family = "binomial")
robust(model)
}
